Refactor: 디렉토리 구조를 DDD로 리팩토링
디렉토리 구조를 DDD로 리팩토링함.

리팩토링한 내용:
- DDD 형태로 변경
