+++ src/main/java/lotto/domain/model/winning/WinningResult.java
@@ -1,4 +1,4 @@
-package lotto.domain.winning;
+package lotto.domain.model.winning;
 
 import java.util.HashMap;
 import java.util.List;
