+++ src/main/java/lotto/exception/winning/WinningNumbersErrorMessages.java
@@ -1,4 +1,4 @@
-package lotto.exception.winningNumbers;
+package lotto.exception.winning;
 
 
 public enum WinningNumbersErrorMessages {
